homework01--understanding of sqa
Software Quality Assurance is a series  planned and organized
activities. It aims at providing  related people with the evidences that
the software has reached the related quality standards. On one hand, SQA
is planned and organised rather than random and arbitrary. Therefore, at
the beginning of a software development, project team should make a plan
for SQA ,guiding he process of software quality assurance activities.On
the other hand, SQA should provides some visibility to the quality of
software products and know whether the software products reach the
appropriate quality standards.whether it follow the appropriate
standards and procedures, and discover the quality problems existing in
the softeware system so that we can find improved method  and measures
to improve the quality of the software.